Modificación en los constructores de las clases y de las interfaces, cambios en la función create ()

+++ src/main/kotlin/models/IMotorGasolina.kt
@@ -2,7 +2,7 @@ package dev.carloscy.models
 
 interface IMotorGasolina: IMotor {
 
-    var consumo: Int
+    var consumo: Double
 
     override fun arrancar() {
         println("Arrancando motor de gasolina.")
